Refactor: Move wishlist logic to a client component
